+++ R/get_delay_estimate.R
@@ -9,17 +9,17 @@
 #'   German Hospitalization Nowcasting Hub.
 #'   Modified from: https://github.com/KITmetricslab/RESPINOW-Hub/blob/7cce3ae2728116e8c8cc0e4ab29074462c24650e/code/baseline/functions.R#L55 #nolint
 #' @param triangle Matrix of the reporting triangle, with rows representing
-#'   the time points of reference and columns representing the delays
+#'   the time points of reference and columns representing the delays.
 #' @param max_delay Integer indicating the maximum delay to estimate, in units
 #'   of the delay. The default is to use the whole reporting triangle,
 #'   `ncol(triangle) -1`.
-#' @param n_history Integer indicating the number of reference dates to be
-#'   used in the estimate of the reporting delay, always starting from the most
-#'   recent reporting delay. The default is to use the whole reporting triangle,
-#'   so `nrow(triangle)-1`
+#' @param n Integer indicating the number of reference times (observations) to
+#'   be used in the estimate of the reporting delay, always starting from the
+#'   most recent reporting delay. The default is to use the whole reporting
+#'   triangle, so `nrow(triangle)`.
 #' @returns Vector indexed at 0 of length `max_delay + 1` with columns
 #'   indicating the point estimate of the empirical probability
-#'   mass on each delay
+#'   mass on each delay.
 #' @export
 #' @examples
 #' triangle <- matrix(
@@ -36,21 +36,21 @@
 #' delay_pmf <- get_delay_estimate(
 #'   triangle = triangle,
 #'   max_delay = 3,
-#'   n_history = 4
+#'   n = 4
 #' )
 #' print(delay_pmf)
 get_delay_estimate <- function(triangle,
                                max_delay = ncol(triangle) - 1,
-                               n_history = nrow(triangle)) {
+                               n = nrow(triangle)) {
   # Check that the input reporting triangle is formatted properly.
   .validate_triangle(triangle,
     max_delay = max_delay,
-    n_history = n_history
+    n = n
   )
 
-  # Filter the triangle down to nrow = n_history + 1, ncol = max_delay
+  # Filter the triangle down to nrow = n_history_delay + 1, ncol = max_delay
   nr0 <- nrow(triangle)
-  trunc_triangle <- triangle[(nr0 - n_history + 1):nr0, 1:(max_delay + 1)]
+  trunc_triangle <- triangle[(nr0 - n + 1):nr0, 1:(max_delay + 1)]
   rep_tri <- .handle_neg_vals(trunc_triangle)
   n_delays <- ncol(rep_tri)
   n_dates <- nrow(rep_tri)