+++ NEWS.md
@@ -1,5 +1,10 @@
 # baselinenowcast 0.0.0.1000
 
+-   Add function to generate a list of expected observed reporting squares
+-   Add function to generate an expected observed reporting square from a point nowcast and a vector of dispersion parameters.
+-   Add function to estimate dispersion parameters from a match list of nowcasts and observed reporting triangles.
+-   Add functions to generate retrospective nowcasts from a single reporting triangle
+-   Refactor uncertainty estimation to use a user-facing function to generate retrospective reporting triangles.
 -   Methods write-up as a separate vignette
 -   Introduced function to estimate the uncertainty from a triangle to be nowcasted and a delay distribution.
 -   Introduced functions to get the delay estimate and apply the delay, and used these in the Getting Started vignette.